# specs/001-python-llm-agent/contracts/api.yaml
openapi: 3.0.3
info:
  title: 多Agent加密货币量化交易分析系统 API
  description: |
    基于Python的多Agent虚拟货币量化交易分析系统API文档

    系统特性：
    - 5个专业agent（新闻收集、做多分析、做空分析、策略生成、交易执行）
    - 支持前5大主流交易所
    - 实时策略分析和自动交易执行
    - 动态资金管理和风险控制
    - 移动端实时监控支持

  version: 1.0.0
  contact:
    name: API Support
    email: support@crypto-ai-trading.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.crypto-ai-trading.com/v1
    description: 生产环境
  - url: https://staging-api.crypto-ai-trading.com/v1
    description: 测试环境
  - url: http://localhost:8000/v1
    description: 开发环境
paths:
...
  /strategies:
    get:
      tags:
        - 交易策略
      summary: 获取策略列表
      description: 获取用户的交易策略列表
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, archived]
          description: 策略状态过滤
        - name: symbol
          in: query
          schema:
            type: string
          description: 交易符号过滤
      responses:
        '200':
          description: 策略列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  strategies:
                    type: array
                    items:
                      $ref: '#/components/schemas/TradingStrategy'

    post:
      tags:
        - 交易策略
      summary: 创建新策略
      description: 创建新的交易策略
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrategyRequest'
      responses:
        '201':
          description: 策略创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingStrategy'
...
  /strategies/{strategy_id}/execute:
    post:
      tags:
        - 交易策略
      summary: 执行策略
      description: 手动触发策略执行
      security:
        - bearerAuth: []
      parameters:
        - name: strategy_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 策略执行结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyExecution'

  # ==================== 交易订单 ====================
  /orders:
    get:
      tags:
        - 交易订单
      summary: 获取订单列表
      description: 获取用户的交易订单列表
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, filled, cancelled, failed]
          description: 订单状态过滤
        - name: symbol
          in: query
          schema:
            type: string
          description: 交易符号过滤
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
          description: 返回条数限制
      responses:
        '200':
          description: 订单列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/TradingOrder'

    post:
      tags:
        - 交易订单
      summary: 创建订单
      description: 创建新的交易订单
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: 订单创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingOrder'
  /orders/{order_id}:
    get:
      tags:
        - 交易订单
      summary: 获取订单详情
      description: 获取指定订单的详细信息
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 订单详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingOrder'

    delete:
      tags:
        - 交易订单
      summary: 取消订单
      description: 取消指定的未完成订单
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 订单取消成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingOrder'

  # ==================== 持仓管理 ====================
  /positions:
    get:
      tags:
        - 持仓管理
      summary: 获取持仓列表
      description: 获取用户的当前持仓信息
      security:
        - bearerAuth: []
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
          description: 交易符号过滤
        - name: side
          in: query
          schema:
            type: string
            enum: [long, short]
          description: 持仓方向过滤
      responses:
        '200':
          description: 持仓列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  positions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Position'

  # ==================== 分析报告 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TradingStrategy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        strategy_type:
          type: string
          enum: [long, short, neutral]
        status:
          type: string
          enum: [active, paused, archived]
        llm_provider:
          type: string
        confidence_score:
          type: number
          format: decimal
        entry_price:
          type: number
          format: decimal
        position_size:
          type: number
          format: decimal
        stop_loss_price:
          type: number
          format: decimal
        take_profit_price:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateStrategyRequest:
      type: object
      required:
        - name
        - symbol
      properties:
        name:
          type: string
        description:
          type: string
        strategy_type:
          type: string
          enum: [long, short, neutral]
        symbol:
          type: string
        entry_price:
          type: number
          format: decimal
        position_size:
          type: number
          format: decimal
        stop_loss_price:
          type: number
          format: decimal
        take_profit_price:
          type: number
          format: decimal
...
    StrategyExecution:
      type: object
      properties:
        execution_id:
          type: string
          format: uuid
        strategy_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        result:
          type: object
        error_message:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time

    # ==================== 交易订单模型 ====================
    TradingOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        order_type:
          type: string
          enum: [market, limit, stop, stop_limit]
        amount:
          type: number
          format: decimal
        price:
          type: number
          format: decimal
        status:
          type: string
          enum: [pending, filled, cancelled, failed]
        filled_amount:
          type: number
          format: decimal
        average_price:
          type: number
          format: decimal
        stop_loss_price:
          type: number
          format: decimal
        take_profit_price:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        filled_at:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required:
        - symbol
        - side
        - order_type
        - amount
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        order_type:
          type: string
          enum: [market, limit, stop, stop_limit]
        amount:
          type: number
          format: decimal
        price:
          type: number
          format: decimal
        stop_loss_price:
          type: number
          format: decimal
        take_profit_price:
          type: number
          format: decimal
        timeout_seconds:
          type: integer
          default: 300

    # ==================== 持仓模型 ====================
    Position:
      type: object
      properties:
        id:
          type: string
          format: uuid
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        amount:
          type: number
          format: decimal
        average_cost:
          type: number
          format: decimal
        current_price:
          type: number
          format: decimal
        unrealized_pnl:
          type: number
          format: decimal
        realized_pnl:
          type: number
          format: decimal
        risk_exposure:
          type: number
          format: decimal
        opened_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ==================== 系统状态模型 ====================
...